+++ rr/dnskey.js
@@ -10,7 +10,7 @@ class DNSKEY extends RR {
   setFlags (val) {
     // a 2 octet Flags Field
     // the possible values are: 0, 256, and 257 RFC 4034
-    if (![ 0, 256, 257 ].includes(val)) throw new Error(`DNSKEY: flags invalid, see ${this.getRFCs()}`)
+    if (![ 0, 256, 257 ].includes(val)) throw new Error(`DNSKEY: flags invalid, see RFC ${this.getRFCs()}`)
 
     this.set('flags', val)
   }
@@ -18,7 +18,7 @@ class DNSKEY extends RR {
   setProtocol (val) {
     // 1 octet
     // The Protocol Field MUST be represented as an unsigned decimal integer with a value of 3.
-    if (![ 3 ].includes(val)) throw new Error(`DNSKEY: protocol invalid, see ${this.getRFCs()}`)
+    if (![ 3 ].includes(val)) throw new Error(`DNSKEY: protocol invalid, see RFC ${this.getRFCs()}`)
 
     this.set('protocol', val)
   }
@@ -27,13 +27,13 @@ class DNSKEY extends RR {
     // 1 octet
     // 1=RSA/MD5, 2=DH, 3=DSA/SHA-1, 4=EC, 5=RSA/SHA-1
     if (![ 1,2,3,4,5,253,254 ].includes(val))
-      throw new Error(`DNSKEY: algorithm invalid, see ${this.getRFCs()}`)
+      throw new Error(`DNSKEY: algorithm invalid, see RFC ${this.getRFCs()}`)
 
     this.set('algorithm', val)
   }
 
   setPublickey (val) {
-    if (!val) throw new Error(`DNSKEY: publickey is required, see ${this.getRFCs()}`)
+    if (!val) throw new Error(`DNSKEY: publickey is required, see RFC ${this.getRFCs()}`)
 
     this.set('publickey', val)
   }